tests: make KgoRepeaterService.await_progress more robust
previously, if you called this right after changing something
in the cluster (e.g. restarting nodes), there was a risk that
you'd enter just before a consumer group fell out of stable state,
and spend the timeout duration of your await_progress just watching
traffic stall while the group got back into a good state.

Now, if await_progress sees that consumers are not progressing, it
will check if the group is in a good state, and if it isn't, then
wait for it to stabilize and reset the timeout.  You only get three
gos at this, to avoid the await_progress potentially running forever.